Feature/ADMIN-5-admin-panel

+++ src/main/java/com/kopchak/worldoftoys/dto/cart/CartItemDto.java
@@ -3,7 +3,8 @@ package com.kopchak.worldoftoys.dto.cart;
 import lombok.Builder;
 
 import java.math.BigDecimal;
+import java.math.BigInteger;
 
 @Builder
-public record CartItemDto(String name, String slug, BigDecimal totalProductPrice, Integer quantity) {
+public record CartItemDto(String name, String slug, BigDecimal totalProductPrice, BigInteger quantity) {
 }
